Init game assets
- model
- view
- presenter
